refactor: 스크롤 초기화 적용과 그를 위한 구조 재정리
